Remove attribute copy mixin and replace with the architectury player event
